+++ lib/whois/record/parser/whois.nic.name.rb
@@ -14,16 +14,14 @@ module Whois
   class Record
     class Parser
 
-      #
-      # = whois.nic.name parser
-      #
       # Parser for the whois.nic.name server.
       #
-      # NOTE: This parser is just a stub and provides only a few basic methods
-      # to check for domain availability and get domain status.
-      # Please consider to contribute implementing missing methods.
-      # See WhoisNicIt parser for an explanation of all available methods
-      # and examples.
+      # @note This parser is just a stub and provides only a few basic methods
+      #   to check for domain availability and get domain status.
+      #   Please consider to contribute implementing missing methods.
+      # 
+      # @see Whois::Record::Parser::Example
+      #   The Example parser for the list of all available methods.
       #
       class WhoisNicName < Base
 
@@ -40,14 +38,30 @@ module Whois
         end
 
 
-        property_not_supported :created_on
+        property_supported :created_on do
+          if content_for_scanner =~ /Created On: (.+)\n/
+            Time.parse($1)
+          end
+        end
 
-        property_not_supported :updated_on
+        property_supported :updated_on do
+          if content_for_scanner =~ /Updated On: (.+)\n/
+            Time.parse($1)
+          end
+        end
 
-        property_not_supported :expires_on
+        property_supported :expires_on do
+          if content_for_scanner =~ /Expires On: (.+)\n/
+            Time.parse($1)
+          end
+        end
 
 
-        property_not_supported :nameservers
+        property_supported :nameservers do
+          content_for_scanner.scan(/Name Server:\s+(.+)\n/).flatten.map do |name|
+            Record::Nameserver.new(:name => name.downcase)
+          end
+        end
 
       end