+++ lib/whois/record/parser/whois.registry.om.rb
@@ -105,7 +105,7 @@ module Whois
         end
 
 
-        private
+      private
 
         def build_contact(element, type)
           node("#{element} Contact ID") do