+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 * SERVER: Added .XN--MGB9AWBF (عمان, Oman) IDN TLD server.
 
+* SERVER: Added .BI TLD server (GH-113).
+
 * NEW: Added full whois.register.com parser.
 
 * NEW: Added full whois.tucows.com parser.
@@ -13,6 +15,8 @@
 * NEW: whois.networksolutions.com parser now recognizes throttled responses (GH-182). [Thanks @JustinCampbell]
 
 * NEW: Added full whois.cmc.iq parser (GH-171).
+
+* NEW: Added full whois1.nic.bi parser (GH-113).
 
 * FIXED: Fixed whois.register.com parser for enom formats (GH-181). [Thanks @JustinCampbell]