+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## master
+## Release 2.7.0
 
 * SERVER: Added .IQ TLD server (GH-171).